only monitor within set times

+++ monitor-with-pir.py
@@ -9,6 +9,12 @@ media_directory = 'home-server/public/images/media'
 def get_file_name():
     return media_directory + '/' + datetime.datetime.now().strftime("%Y-%m-%d_%H.%M.%S")
 
+def in_hours():
+    now = datetime.datetime.now()
+    weekday = now.weekday()
+    hour = now.hour
+    return weekday !== 5 and weekday !== 6 and hour > 9 and hour < 18
+
 sensor = 4
 
 try:
@@ -26,22 +32,26 @@ with picamera.PiCamera() as cam:
     cam.resolution = (800, 600)
     print("Starting at " + str(datetime.datetime.now()))
     while True:
-        time.sleep(0.1)
-        previous_state = current_state
-        current_state = GPIO.input(sensor)
-        if current_state != previous_state:
-            new_state = "HIGH" if current_state else "LOW"
-            print("GPIO pin %s is %s" % (sensor, new_state))
-            if current_state:
-                fileName = get_file_name()
-                print("starting recording at "+fileName)
-                cam.start_recording(fileName + '.h264')
-                cam.wait_recording(10)
-                cam.capture(fileName+'.jpg', use_video_port=True)
-                cam.wait_recording(20)
-                cam.stop_recording()
-                print("stopped recording "+fileName)
-                print("converting to mp4")
-                os.system("MP4Box -fps 30 -add %s.h264 %s.mp4" % (fileName, fileName))
-                os.remove(fileName + '.h264')
-                print("\a") # ring bell in terminal
+        if in_hours():
+            time.sleep(0.1)
+            previous_state = current_state
+            current_state = GPIO.input(sensor)
+            if current_state != previous_state:
+                new_state = "HIGH" if current_state else "LOW"
+                print("GPIO pin %s is %s" % (sensor, new_state))
+                if current_state:
+                    fileName = get_file_name()
+                    print("starting recording at "+fileName)
+                    cam.start_recording(fileName + '.h264')
+                    cam.wait_recording(10)
+                    cam.capture(fileName+'.jpg', use_video_port=True)
+                    cam.wait_recording(20)
+                    cam.stop_recording()
+                    print("stopped recording "+fileName)
+                    print("converting to mp4")
+                    os.system("MP4Box -fps 30 -add %s.h264 %s.mp4" % (fileName, fileName))
+                    os.remove(fileName + '.h264')
+                    print("\a") # ring bell in terminal
+        else:
+            time.sleep(60)
+            print(".")
